+++ xmtp_api_http/Cargo.toml
@@ -39,6 +39,7 @@ xmtp_proto = { path = "../xmtp_proto", features = ["test-utils"] }
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread", "time"] }
 ctor.workspace = true
+httpmock = "0.7"
 
 [target.'cfg(target_arch = "wasm32")'.dev-dependencies]
 tokio = { workspace = true, features = ["macros", "time"] }
