+++ xmtp_api_http/src/test.rs
@@ -1,0 +1,4 @@
+pub mod mock;
+
+#[cfg(not(target_arch = "wasm32"))]
+mod native_tests;
